Add acc data management to atm_recover_large_step_variables_work
These changes ensure that the other, non-invariant, fields are available
on the device during this routine. Some fields that are overwritten are
only created at the beginning, while others are copied in. Any fields
that were unmodified right-hand side fields are deleted at the end and
modified left-hand side fields are copied out.

Timing for these transfers are reported in the output log file in the
new timer: `atm_recover_large_step_variables [ACC_data_xfer]`.

'default(present)' clauses are added to the parallel directives so that
any missing data on the device causes a runtime error.