Disable physics for MPAS-A dycore-only build
When building MPAS-A as a dycore, all physics-related components are disabled.
This build configuration is usually used by CAM/CAM-SIMA, and can be achieved
by defining the `MPAS_CAM_DYCORE` macro in `CPPFLAGS`.

The `PHYSICS` variable controls whether physics are enabled in MPAS-A, but its
logic is decoupled from the `MPAS_CAM_DYCORE` macro. Disabling physics in MPAS-A
currently requires manual interventions.

Therefore, automatically disable physics when the `MPAS_CAM_DYCORE` macro is
found in `CPPFLAGS`.